refactor: BE - JWT payload builder 패턴 적용
