+++ backend/src/auth/dto/create-jwt.dto.ts
@@ -1,6 +1,9 @@
 import { IsEmail, IsString } from 'class-validator';
 
-export class CreateJwtDto {
+export class CreateJwtPayloadDto {
+	@IsString()
+	id: string;
+
 	@IsString()
 	nickname: string;
